+++ src/main/java/act/cli/view/CliView.java
@@ -9,9 +9,9 @@ package act.cli.view;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,25 +20,21 @@ package act.cli.view;
  * #L%
  */
 
-import act.Act;
 import act.cli.CliContext;
 import act.cli.CliOverHttpContext;
 import act.cli.ascii_table.impl.CollectionASCIITableAware;
 import act.cli.tree.TreeNode;
 import act.cli.util.CliCursor;
-import act.cli.util.MappedFastJsonNameFilter;
 import act.cli.util.TableCursor;
 import act.data.DataPropertyRepository;
 import act.db.AdaptiveRecord;
 import act.util.ActContext;
-import act.util.DisableFastJsonCircularReferenceDetect;
-import act.util.FastJsonPropertyPreFilter;
+import act.util.JsonUtilConfig;
 import act.util.PropertySpec;
-import com.alibaba.fastjson.serializer.SerializeFilter;
-import com.alibaba.fastjson.serializer.SerializerFeature;
 import org.osgl.$;
 import org.osgl.util.C;
 import org.osgl.util.E;
+import org.osgl.util.Output;
 import org.osgl.util.S;
 import org.rythmengine.utils.Escape;
 
@@ -55,10 +51,10 @@ public enum CliView {
     TABLE() {
         @Override
         @SuppressWarnings("unchecked")
-        public String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
+        public void render(Output output, Object result, PropertySpec.MetaInfo spec, ActContext context) {
 
             if (null == result) {
-                return "no data";
+                return;
             }
 
             spec = PropertySpec.MetaInfo.withCurrent(spec, context);
@@ -79,11 +75,11 @@ public enum CliView {
                 TableCursor cursor = new TableCursor(dataList, pageSize, spec);
                 cliContext.session().cursor(cursor);
                 cursor.output(cliContext);
-                return "";
+                return;
             }
             Class<?> componentType;
             if (dataList.isEmpty()) {
-                return "no data";
+                return;
             }
             componentType = Object.class;
             for (Object o : dataList) {
@@ -105,7 +101,7 @@ public enum CliView {
                 itemsFound = cursor.records();
                 appendix = cursor.hasNext() ? "\nType \"it\" for more" : "";
             }
-            return S.concat(tableString, "Items found: ", S.string(itemsFound), appendix);
+            output.append(S.concat(tableString, "Items found: ", S.string(itemsFound), appendix));
         }
 
     },
@@ -123,27 +119,25 @@ public enum CliView {
      */
     TREE() {
         @Override
-        public String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
+        public void render(Output output, Object result, PropertySpec.MetaInfo spec, ActContext context) {
             if (result instanceof TreeNode) {
-                return toTreeString((TreeNode) result);
+                toTreeString(output, (TreeNode) result);
             } else if (result instanceof Iterable) {
-                return TABLE.render(result, spec, context);
+                TABLE.render(output, result, spec, context);
             } else if (null != spec) {
-                return JSON.render(result, spec, context);
+                JSON.render(output, result, spec, context);
             } else {
-                return S.string(result);
+                output.append(S.string(result));
             }
         }
 
-        private String toTreeString(TreeNode result) {
-            StringBuilder sb = S.newBuilder();
-            buildTree(sb, result, "", true);
-            return sb.toString();
+        private void toTreeString(Output output, TreeNode result) {
+            buildTree(output, result, "", true);
         }
 
-        private void buildTree(StringBuilder sb, TreeNode node, String prefix, boolean isTrail) {
-            StringBuilder sb0 = S.newBuilder().append(prefix).append(isTrail ? "└── " : "├── ").append(node.label()).append("\n");
-            sb.append(sb0);
+        private void buildTree(Output output, TreeNode node, String prefix, boolean isTrail) {
+            StringBuilder sb = S.newBuilder().append(prefix).append(isTrail ? "└── " : "├── ").append(node.label()).append("\n");
+            output.append(sb);
             List<TreeNode> children = node.children();
             int sz = children.size();
             if (sz == 0) {
@@ -152,9 +146,9 @@ public enum CliView {
             final String subPrefix = S.newBuilder().append(prefix).append(isTrail ? "    " : "│   ").toString();
             for (int i = 0; i < sz - 1; ++i) {
                 TreeNode child = children.get(i);
-                buildTree(sb, child, subPrefix, false);
+                buildTree(output, child, subPrefix, false);
             }
-            buildTree(sb, children.get(sz - 1), subPrefix, true);
+            buildTree(output, children.get(sz - 1), subPrefix, true);
         }
     },
 
@@ -163,93 +157,23 @@ public enum CliView {
      */
     XML() {
         @Override
-        public String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
+        public void render(Output output, Object result, PropertySpec.MetaInfo spec, ActContext context) {
             throw E.unsupport();
         }
     },
 
     /**
-     * Present the result using {@link act.cli.JsonView}
+     * Present the result using {@link act.util.JsonView}
      */
     JSON() {
         @Override
-        public String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
-            return render(result, spec, context, context instanceof CliContext);
+        public void render(Output output, Object result, PropertySpec.MetaInfo spec, ActContext context) {
+            render(output, result, spec, context, context instanceof CliContext);
         }
 
-        public String render(Object result, PropertySpec.MetaInfo spec, ActContext context, boolean format) {
-            String json;
-            FastJsonPropertyPreFilter propertyFilter;
-            spec = PropertySpec.MetaInfo.withCurrent(spec, context);
-            if (null == spec) {
-                propertyFilter = null;
-            } else {
-                propertyFilter = new FastJsonPropertyPreFilter();
-                List<String> outputs = spec.outputFields(context);
-                Set<String> excluded = spec.excludedFields(context);
-                if (excluded.isEmpty()) {
-                    if (outputs.isEmpty()) {
-                        propertyFilter = null; // no filter defined actually
-                    } else {
-                        // output fields only applied when excluded fields not presented
-                        propertyFilter.addIncludes(outputs);
-                        if (FastJsonPropertyPreFilter.hasPattern(outputs)) {
-                            // TODO: handle the case when result is an Iterable
-                            propertyFilter.setFullPaths(context.app().service(DataPropertyRepository.class).propertyListOf(result.getClass()));
-                        }
-                    }
-                } else {
-                    propertyFilter.addExcludes(excluded);
-                    if (FastJsonPropertyPreFilter.hasPattern(excluded)) {
-                        // TODO: handle the case when result is an Iterable
-                        propertyFilter.setFullPaths(context.app().service(DataPropertyRepository.class).propertyListOf(result.getClass()));
-                    }
-                }
-            }
-
-            List<SerializerFeature> featureList = new ArrayList<>();
-            if (format) {
-                featureList.add(SerializerFeature.PrettyFormat);
-            }
-            if (null == propertyFilter) {
-                Boolean b = DisableFastJsonCircularReferenceDetect.option.get();
-                if (null != b && b) {
-                    featureList.add(SerializerFeature.DisableCircularReferenceDetect);
-                }
-                SerializerFeature[] featureArray = new SerializerFeature[featureList.size()];
-                featureArray = featureList.toArray(featureArray);
-                if (format) {
-                    json = com.alibaba.fastjson.JSON.toJSONString(result, featureArray);
-                } else {
-                    json = com.alibaba.fastjson.JSON.toJSONString(result);
-                }
-            } else {
-                Boolean b = DisableFastJsonCircularReferenceDetect.option.get();
-                if (null != b && b) {
-                    Act.LOGGER.warn(new RuntimeException(), "Cannot use @DisableFastJsonCircularReferenceDetect along with @PropertySpec");
-                    // Note: we can't check DisableFastJsonCircularReferenceDetect here because if
-                    // that option is set, then FastJson will skip the JsonSerializer.context setting
-                    // and breaks the property filter mechanism
-                    //featureList.add(SerializerFeature.DisableCircularReferenceDetect);
-                }
-                MappedFastJsonNameFilter nameFilter = new MappedFastJsonNameFilter(spec.labelMapping(context));
-
-                SerializerFeature[] featureArray = new SerializerFeature[featureList.size()];
-                featureArray = featureList.toArray(featureArray);
-
-                if (nameFilter.isEmpty()) {
-                    json = com.alibaba.fastjson.JSON.toJSONString(result, propertyFilter, featureArray);
-                } else {
-                    SerializeFilter[] filters = new SerializeFilter[2];
-                    filters[0] = nameFilter;
-                    filters[1] = propertyFilter;
-                    json = com.alibaba.fastjson.JSON.toJSONString(result, filters, featureArray);
-                }
-            }
-            return json;
+        public void render(Output output, Object result, PropertySpec.MetaInfo spec, ActContext context, boolean format) {
+            new JsonUtilConfig.JsonWriter(result, spec, format, context).apply(output);
         }
-
-
     },
 
     /**
@@ -257,40 +181,57 @@ public enum CliView {
      */
     TO_STRING() {
         @Override
-        public String render(Object result, PropertySpec.MetaInfo filter, ActContext context) {
+        public void render(Output output, Object result, PropertySpec.MetaInfo filter, ActContext context) {
             if (result instanceof Iterable) {
-                return TABLE.render(result, filter, context);
+                TABLE.render(output, result, filter, context);
             } else if (result instanceof TreeNode) {
-                return TREE.render(result, filter, context);
+                TREE.render(output, result, filter, context);
             } else if (null != filter) {
-                return JSON.render(result, filter, context);
+                JSON.render(output, result, filter, context);
             } else {
-                return S.string(result);
+                output.append(S.string(result));
             }
         }
     },
 
     CSV() {
         @Override
-        public String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
+        public void render(Output output, Object result, PropertySpec.MetaInfo spec, ActContext context) {
             if (null == result) {
-                return "no data";
+                return;
             }
-            List dataList;
+            Iterator iterator;
             Class<?> componentType;
             if (result instanceof Iterable) {
-                dataList = C.list((Iterable) result);
+                iterator = ((Iterable) result).iterator();
             } else if (result instanceof Iterator) {
-                dataList = C.list((Iterator) result);
+                iterator = (Iterator) result;
             } else if (result instanceof Enumeration) {
-                dataList = C.list((Enumeration) result);
+                final Enumeration enumeration = (Enumeration) result;
+                iterator = new Iterator() {
+                    @Override
+                    public boolean hasNext() {
+                        return enumeration.hasMoreElements();
+                    }
+
+                    @Override
+                    public Object next() {
+                        return enumeration.nextElement();
+                    }
+
+                    @Override
+                    public void remove() {
+                        throw E.unsupport();
+                    }
+                };
             } else {
-                dataList = C.list(result);
+                iterator = C.list(result).iterator();
             }
-            if (dataList.isEmpty()) {
-                return "no data";
+            Object firstElement = iterator.hasNext() ? iterator.next() : null;
+            if (null == firstElement) {
+                return;
             }
-            componentType = dataList.get(0).getClass();
+            componentType = firstElement.getClass();
             DataPropertyRepository repo = context.app().service(DataPropertyRepository.class);
             spec = PropertySpec.MetaInfo.withCurrent(spec, context);
             if (null == spec) {
@@ -298,22 +239,25 @@ public enum CliView {
                 spec.onValue("-not_exists");
             }
             List<String> outputFields = repo.outputFields(spec, componentType, context);
-            S.Buffer sb = S.buffer();
-            buildHeaderLine(sb, outputFields, spec.labelMapping());
-            for (Object entity : dataList) {
-                sb.append($.OS.lineSeparator());
-                buildDataLine(sb, entity, outputFields);
+            output.append(buildHeaderLine(outputFields, spec.labelMapping()));
+            output.append($.OS.lineSeparator());
+            output.append(buildDataLine(firstElement, outputFields));
+            while (iterator.hasNext()) {
+                output.append($.OS.lineSeparator());
+                output.append(buildDataLine(iterator.next(), outputFields));
+                output.flush();
             }
-            return sb.toString();
         }
 
-        private void buildDataLine(S.Buffer sb, Object data, List<String> outputFields) {
+        private String buildDataLine(Object data, List<String> outputFields) {
             Iterator<String> itr = outputFields.iterator();
             String prop = itr.next();
-            sb.append(getProperty(data, prop));
+            S.Buffer buf = S.buffer();
+            buf.append(getProperty(data, prop));
             while (itr.hasNext()) {
-                sb.append(",").append(getProperty(data, itr.next()));
+                buf.append(",").append(getProperty(data, itr.next()));
             }
+            return buf.toString();
         }
 
         private String getProperty(Object data, String prop) {
@@ -327,16 +271,18 @@ public enum CliView {
             }
         }
 
-        private void buildHeaderLine(S.Buffer sb, List<String> outputFields, Map<String, String> labels) {
+        private String buildHeaderLine(List<String> outputFields, Map<String, String> labels) {
             if (null == labels) {
                 labels = new HashMap<>();
             }
             Iterator<String> itr = outputFields.iterator();
             String label = label(itr.next(), labels);
-            sb.append(label);
+            S.Buffer buf = S.buffer();
+            buf.append(label);
             while (itr.hasNext()) {
-                sb.append(",").append(escape(label(itr.next(), labels)));
+                buf.append(",").append(escape(label(itr.next(), labels)));
             }
+            return buf.toString();
         }
 
         private String label(String key, Map<String, String> labels) {
@@ -345,13 +291,26 @@ public enum CliView {
         }
 
         private String escape(Object o) {
+            if (null == o) {
+                return "";
+            }
+            String s = o.toString().trim();
+            if (s.startsWith("\"") && s.endsWith("\"")) {
+                return s;
+            }
             return Escape.CSV.apply(o).toString();
         }
 
     };
 
-    public String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
+    public void render(Output out, Object result, PropertySpec.MetaInfo spec, ActContext context) {
         throw E.unsupport();
+    }
+
+    public final String render(Object result, PropertySpec.MetaInfo spec, ActContext context) {
+        S.Buffer sb = S.buffer();
+        render(sb, result, spec, context);
+        return sb.toString();
     }
 
     public void print(Object result, PropertySpec.MetaInfo spec, CliContext context) {