+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **1.7.4** 
 * notFoundIfNull return Content-Type as text/html #537
+* Implement new response output model #539
 
 **1.7.3**
 * Fix regression issue with #504